requirements.txt
-f /usr/share/pip-wheels
annotator==0.14.2
click==6.6
Django==1.10.4
django-filter==1.0.1
djangorestframework==3.5.3
elasticsearch==1.9.0
Flask==0.11.1
iso8601==0.1.11
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mysqlclient==1.3.9
PyJWT==1.4.2
python-dateutil==2.6.0
pytz==2016.10
regex==2016.12.27
requests==2.11.1
six==1.10.0
urllib3==1.19.1
Werkzeug==0.11.11
